# Optimize listener generation