--- a/pilot/pkg/networking/core/v1alpha3/listener.go
+++ b/pilot/pkg/networking/core/v1alpha3/listener.go
@@ -58,7 +58,6 @@ import (
 	"istio.io/istio/pkg/config/protocol"
 	"istio.io/istio/pkg/proto"
 	"istio.io/istio/pkg/util/gogo"
-	alpn_filter "istio.io/istio/security/proto/envoy/config/filter/http/alpn/v2alpha1"
 )
 
 const (
@@ -1114,14 +1113,7 @@ func (configgen *ConfigGeneratorImpl) buildSidecarOutboundListeners(node *model.
 	// Now validate all the listeners. Collate the tcp listeners first and then the HTTP listeners
 	// TODO: This is going to be bad for caching as the order of listeners in tcpListeners or httpListeners is not
 	// guaranteed.
-	invalid := 0.0
-	for name, l := range listenerMap {
-		if err := l.listener.Validate(); err != nil {
-			log.Warnf("buildSidecarOutboundListeners: error validating listener %s (type %v): %v", name, l.servicePort.Protocol, err)
-			invalid++
-			invalidOutboundListeners.Record(invalid)
-			continue
-		}
+	for _, l := range listenerMap {
 		if l.servicePort.Protocol.IsTCP() {
 			tcpListeners = append(tcpListeners, l.listener)
 		} else {
@@ -1479,10 +1471,14 @@ func (configgen *ConfigGeneratorImpl) buildSidecarOutboundListenerForPortOrUDS(n
 				// append the service. Otherwise (TCP), change current listener to use protocol sniffing.
 				if currentListenerEntry.protocol.IsHTTP() {
 					conflictType = HTTPOverHTTP
+					currentListenerEntry.services = append(currentListenerEntry.services, pluginParams.Service)
+					return
 				} else if currentListenerEntry.protocol.IsTCP() {
 					conflictType = HTTPOverTCP
 				} else {
 					conflictType = HTTPOverAuto
+					currentListenerEntry.services = append(currentListenerEntry.services, pluginParams.Service)
+					return
 				}
 			}
 
@@ -1580,6 +1576,8 @@ func (configgen *ConfigGeneratorImpl) buildSidecarOutboundListenerForPortOrUDS(n
 					conflictType = AutoOverTCP
 				} else {
 					conflictType = AutoOverAuto
+					currentListenerEntry.services = append(currentListenerEntry.services, pluginParams.Service)
+					return
 				}
 			}
 
@@ -1658,10 +1656,6 @@ func (configgen *ConfigGeneratorImpl) buildSidecarOutboundListenerForPortOrUDS(n
 				protocol:    pluginParams.Port.Protocol,
 			}
 		}
-
-	case HTTPOverHTTP, HTTPOverAuto:
-		// Append service only
-		currentListenerEntry.services = append(currentListenerEntry.services, pluginParams.Service)
 	case HTTPOverTCP:
 		// Merge HTTP filter chain to TCP filter chain
 		currentListenerEntry.listener.FilterChains = mergeFilterChains(mutable.Listener.FilterChains, currentListenerEntry.listener.FilterChains)
@@ -1703,8 +1697,10 @@ func (configgen *ConfigGeneratorImpl) buildSidecarOutboundListenerForPortOrUDS(n
 		currentListenerEntry.protocol = protocol.Unsupported
 		currentListenerEntry.listener.ListenerFilters = appendListenerFilters(currentListenerEntry.listener.ListenerFilters)
 
-	case AutoOverAuto:
-		currentListenerEntry.services = append(currentListenerEntry.services, pluginParams.Service)
+	default:
+		// Covered previously - in this case we return early to prevent creating listeners that we end up throwing away
+		// This should never happen
+		log.Errorf("Got unexpected conflict type %v. This should never happen", conflictType)
 	}
 
 	if log.DebugEnabled() && len(mutable.Listener.FilterChains) > 1 || currentListenerEntry != nil {
@@ -1916,27 +1912,7 @@ func buildHTTPConnectionManager(pluginParams *plugin.InputParams, httpOpts *http
 
 	// append ALPN HTTP filter in HTTP connection manager for outbound listener only.
 	if pluginParams.ListenerCategory == networking.EnvoyFilter_SIDECAR_OUTBOUND {
-		filters = append(filters, &http_conn.HttpFilter{
-			Name: AlpnFilterName,
-			ConfigType: &http_conn.HttpFilter_TypedConfig{
-				TypedConfig: util.MessageToAny(&alpn_filter.FilterConfig{
-					AlpnOverride: []*alpn_filter.FilterConfig_AlpnOverride{
-						{
-							UpstreamProtocol: alpn_filter.FilterConfig_HTTP10,
-							AlpnOverride:     mtlsHTTP10ALPN,
-						},
-						{
-							UpstreamProtocol: alpn_filter.FilterConfig_HTTP11,
-							AlpnOverride:     mtlsHTTP11ALPN,
-						},
-						{
-							UpstreamProtocol: alpn_filter.FilterConfig_HTTP2,
-							AlpnOverride:     mtlsHTTP2ALPN,
-						},
-					},
-				}),
-			},
-		})
+		filters = append(filters, alpnFilter)
 	}
 
 	filters = append(filters, corsFilter, faultFilter, routerFilter)