This has a couple small changes:
* Fix to the FakeStore to allow benchmark to run. We don't (yet!) run
this in CI
* Premarshall the ALPN filter to avoid creating it 100s of times
* If we have a conflict, short circuit to avoid construction a full
filter chain and throwing it away
* Stop calling listener.Validate() (see
https://github.com/istio/istio/issues/22648 for more info)

```
name                                 old time/op    new time/op    delta
ListenerGeneration/empty-6             4.09ms ± 3%    2.92ms ± 0%  -28.70%  (p=0.008 n=5+5)
ListenerGeneration/telemetry-6         5.00ms ± 4%    3.88ms ± 2%  -22.39%  (p=0.008 n=5+5)
ListenerGeneration/virtualservice-6    4.05ms ± 2%    2.92ms ± 2%  -27.88%  (p=0.008 n=5+5)

name                                 old alloc/op   new alloc/op   delta
ListenerGeneration/empty-6             1.81MB ± 0%    1.35MB ± 0%  -25.61%  (p=0.008 n=5+5)
ListenerGeneration/telemetry-6         2.12MB ± 0%    1.66MB ± 0%  -21.79%  (p=0.008 n=5+5)
ListenerGeneration/virtualservice-6    1.81MB ± 0%    1.35MB ± 0%  -25.60%  (p=0.008 n=5+5)

name                                 old allocs/op  new allocs/op  delta
ListenerGeneration/empty-6              26.9k ± 0%     19.6k ± 0%  -27.28%  (p=0.008 n=5+5)
ListenerGeneration/telemetry-6          32.3k ± 0%     25.0k ± 0%     ~     (p=0.079 n=4+5)
ListenerGeneration/virtualservice-6     26.9k ± 0%     19.6k ± 0%     ~     (p=0.079 n=4+5)
```